158_tokenizer/requirements.txt
PyICU
mecab-python3
jsonrpcserver>=4,<5
jsonrpcclient[requests]
werkzeug
uwsgi
